+++ test/server/src/parser.test.ts
@@ -2655,6 +2655,22 @@ describe("Parser", () => {
 				expect(received[0].range.end.line).to.equal(7);
 			});
 
+			it("should flag an operator in a bare multireplace even without spaces", () => {
+				let fakeDocument = createDocument("@{var>2 yes | no}");
+				let received: Array<Diagnostic> = [];
+				let fakeCallbacks = Substitute.for<ParserCallbacks>();
+				fakeCallbacks.onParseError(Arg.all()).mimicks((e: Diagnostic) => {
+					received.push(e);
+				});
+
+				parse(fakeDocument, fakeCallbacks);
+
+				expect(received.length).to.equal(1);
+				expect(received[0].message).to.include("Potentially missing parentheses");
+				expect(received[0].range.start.line).to.equal(2);
+				expect(received[0].range.end.line).to.equal(7);
+			});
+
 			it("should flag an operator in a bare multireplace inside a block", () => {
 				let fakeDocument = createDocument("*if true\n\t@{var > 2 yes | no}");
 				let received: Array<Diagnostic> = [];